+++ nuorder/__main__.py
@@ -1,11 +1,39 @@
-import termcolor
+"""Package for making HTTP requests to the NuOrder API
+
+It's recommended to put hostname and consumer/oauth settings in an INI-style
+file at the location ~/.config/nuorder.ini. This way you don't have to
+put them in for every request.
+
+        [sandbox]
+        hostname = wholesale.sandbox1.nuorder.com
+        consumer_key = QDaGd4ppfXTPEaxnjz4C
+        consumer_secret = ZvbKP5jxL0iBJ2p7zNRsBzG9vo8XdSIVLb1fMkWFX55dsKTL
+        oauth_token = 74SCldgh0DfBufxKJTlEe
+        oauth_token_secret = Eb6haktmLIeTYO0LuyCktJNADpYPMnvo6rWWKOs6oh1WJH
+        app_name = My app
+
+Then you can simply issue a request to their API like this::
+
+    nuorder get /api/companies/codes/list
+
+Please note that the config section name `sandbox` is the
+default and can be overridden by passing in `-c <name>`.
+Useful when one wants to make calls to both sandbox and production
+environments.
+"""
+import functools
 import json
 import logging
 import sys
+import termcolor
 
+import IPython
 import argh
+import requests
+import traitlets
 
 import nuorder
+from nuorder import ini_config
 
 
 def _failure(err):
@@ -13,33 +41,99 @@ def _failure(err):
 
 
 def _set_log_level(log_level):
-    nuorder.logger.setLevel(getattr(logging, log_level.upper()))
+    if log_level is not None:
+        level = getattr(logging, log_level.upper())
+        logging.basicConfig()
+        nuorder.logger.setLevel(level)
 
 
-@argh.wrap_errors([nuorder.config.ConfigKeyMissing], processor=_failure)
+DEFAULT_WRAPPED_ERRORS = [
+    ini_config.ConfigKeyMissing,
+    requests.exceptions.HTTPError,
+]
+
+
+@argh.wrap_errors([ini_config.ConfigKeyMissing], processor=_failure)
 @argh.arg('-c', '--config-section')
+def interact(
+    hostname: 'E.g. wholesale.sandbox1.nuorder.com for sandbox' = None,
+    consumer_key: 'The consumer key to use' = None,
+    consumer_secret: 'The oauth shared secret to use' = None,
+    oauth_token: 'OAuth token' = None,
+    oauth_token_secret: 'OAuth token secret' = None,
+    config_section: 'The name of the config section to get settings from.' = 'sandbox',
+    log_level: 'The log level to use.' = None,
+):
+    """
+    Starts an IPython REPL to enable easy interaction with the NuOrder API
+    """
+    _set_log_level(log_level)
+    c = functools.partial(ini_config.get, config_section)
+
+    nu = nuorder.NuOrder(
+        hostname=hostname or c('hostname'),
+        consumer_key=consumer_key or c('consumer_key'),
+        consumer_secret=consumer_secret or c('consumer_secret'),
+        oauth_token=oauth_token or c('oauth_token'),
+        oauth_token_secret=oauth_token_secret or c('oauth_token_secret'),
+    )
+
+    banner1 = """Welcome to the `nuorder` client interactive mode
+Available vars:
+    {additional_arg}
+    `nuorder` - An instantiation of the nuorder.NuOrder class
+
+Example usage:
+    schemas = nuorder.get('/api/schemas')
+    products = nuorder.get('/api/products')
+    {additional_arg_example}
+
+Notes:
+    * Official NuOrder docs can be found online here:
+      https://nuorderapi1.docs.apiary.io/
+    * Creation of oauth tokens can't be done inside this mode. Please see `nuorder initiate --help`.
+"""
+    IPython.embed(
+        user_ns={'nuorder': nu},
+        banner1=banner1,
+        config=traitlets.config.Config(colors='LightBG')
+    )
+
+
+@argh.wrap_errors(DEFAULT_WRAPPED_ERRORS, processor=_failure)
+@argh.arg('-c', '--config-section')
+@argh.arg('--dry-run')
 def initiate(
     hostname: 'E.g. wholesale.sandbox1.nuorder.com for sandbox' = None,
     consumer_key: 'The consumer key to use' = None,
     consumer_secret: 'The oauth shared secret to use' = None,
     app_name: 'The application name.' = None,
     config_section: 'The name of the config section to get settings from.' = 'sandbox',
-    log_level: 'The log level to use.' = 'WARNING',
+    log_level: 'The log level to use.' = None,
     dry_run: "Don't actually run command, just show what would be run." = False
 ):
-    """Make GET requests to NuOrder"""
+    """Generate a new oauth token and secret
+
+    Interactive command that requires you to go to the NuOrder admin
+    page and confirm the request.
+    """
     _set_log_level(log_level)
 
-    request_args = dict(
-        method='GET',
-        hostname=hostname,
-        consumer_key=consumer_key,
-        consumer_secret=consumer_secret,
-        config_section=config_section,
-        app_name=app_name,
-        dry_run=dry_run,
+    c = functools.partial(ini_config.get, config_section)
+
+    base_kwargs = {
+        'hostname': hostname or c('hostname'),
+        'consumer_key': consumer_key or c('consumer_key'),
+        'consumer_secret': consumer_secret or c('consumer_secret'),
+    }
+
+    nu = nuorder.NuOrder(
+        **base_kwargs,
+        oauth_token='',
+        oauth_token_secret='',
     )
-    resp_json = nuorder.request(endpoint='/api/initiate', **request_args)
+
+    resp_json = nu.oauth_initiate(app_name or c('app_name'), dry_run=dry_run)
     resp_text = json.dumps(resp_json, indent=2)
 
     if 'request_error' in resp_json:
@@ -59,13 +153,12 @@ def initiate(
     )
     verifier = input("Verification code [paste and press Enter]: ")
 
-    resp_json = nuorder.request(
-        endpoint='/api/token',
-        **request_args,
+    nu = nuorder.NuOrder(
+        **base_kwargs,
         oauth_token=resp_json['oauth_token'],
         oauth_token_secret=resp_json['oauth_token_secret'],
-        oauth_verifier=verifier,
     )
+    resp_json = nu.oauth_token_request(verifier, dry_run=dry_run)
 
     if 'request_error' in resp_json:
         sys.exit(_failure(json.dumps(resp_json, indent=2)))
@@ -78,8 +171,9 @@ def initiate(
     return json.dumps(resp_json, indent=2)
 
 
-@argh.wrap_errors([nuorder.config.ConfigKeyMissing], processor=_failure)
+@argh.wrap_errors(DEFAULT_WRAPPED_ERRORS, processor=_failure)
 @argh.arg('-c', '--config-section')
+@argh.arg('--dry-run')
 def delete(
     endpoint: 'The endpoint to interact with.',
     hostname: 'E.g. wholesale.sandbox1.nuorder.com for sandbox' = None,
@@ -88,27 +182,28 @@ def delete(
     oauth_token: 'OAuth token' = None,
     oauth_token_secret: 'OAuth token secret' = None,
     config_section: 'The name of the config section to get settings from.' = 'sandbox',
-    log_level: 'The log level to use.' = 'WARNING',
+    log_level: 'The log level to use.' = None,
     dry_run: "Don't actually run command, just show what would be run." = False
 ):
     """Make a DELETE request to NuOrder"""
     _set_log_level(log_level)
-    resp_json = nuorder.request(
-        method='DELETE',
-        endpoint=endpoint,
-        hostname=hostname,
-        consumer_key=consumer_key,
-        consumer_secret=consumer_secret,
-        oauth_token=oauth_token,
-        oauth_token_secret=oauth_token_secret,
-        config_section=config_section,
-        dry_run=dry_run,
+
+    c = functools.partial(ini_config.get, config_section)
+
+    nu = nuorder.NuOrder(
+        hostname=hostname or c('hostname'),
+        consumer_key=consumer_key or c('consumer_key'),
+        consumer_secret=consumer_secret or c('consumer_secret'),
+        oauth_token=oauth_token or c('oauth_token'),
+        oauth_token_secret=oauth_token_secret or c('oauth_token_secret'),
     )
+    resp_json = nu.delete(endpoint, dry_run=dry_run)
     return json.dumps(resp_json, indent=2)
 
 
-@argh.wrap_errors([nuorder.config.ConfigKeyMissing], processor=_failure)
+@argh.wrap_errors(DEFAULT_WRAPPED_ERRORS, processor=_failure)
 @argh.arg('-c', '--config-section')
+@argh.arg('--dry-run')
 def get(
     endpoint: 'The endpoint to interact with.',
     hostname: 'E.g. wholesale.sandbox1.nuorder.com for sandbox' = None,
@@ -117,27 +212,29 @@ def get(
     oauth_token: 'OAuth token' = None,
     oauth_token_secret: 'OAuth token secret' = None,
     config_section: 'The name of the config section to get settings from.' = 'sandbox',
-    log_level: 'The log level to use.' = 'WARNING',
+    log_level: 'The log level to use.' = None,
     dry_run: "Don't actually run command, just show what would be run." = False
 ):
     """Make a GET request to NuOrder"""
     _set_log_level(log_level)
-    resp_json = nuorder.request(
-        method='GET',
-        endpoint=endpoint,
-        hostname=hostname,
-        consumer_key=consumer_key,
-        consumer_secret=consumer_secret,
-        oauth_token=oauth_token,
-        oauth_token_secret=oauth_token_secret,
-        config_section=config_section,
-        dry_run=dry_run,
+
+    c = functools.partial(ini_config.get, config_section)
+
+    nu = nuorder.NuOrder(
+        hostname=hostname or c('hostname'),
+        consumer_key=consumer_key or c('consumer_key'),
+        consumer_secret=consumer_secret or c('consumer_secret'),
+        oauth_token=oauth_token or c('oauth_token'),
+        oauth_token_secret=oauth_token_secret or c('oauth_token_secret'),
     )
+    resp_json = nu.get(endpoint=endpoint, dry_run=dry_run)
     return json.dumps(resp_json, indent=2)
 
 
-@argh.wrap_errors([nuorder.config.ConfigKeyMissing], processor=_failure)
+@argh.wrap_errors(DEFAULT_WRAPPED_ERRORS, processor=_failure)
 @argh.arg('-c', '--config-section')
+@argh.arg('-d', '--data')
+@argh.arg('--dry-run')
 def post(
     endpoint: 'The endpoint to interact with.',
     hostname: 'E.g. wholesale.sandbox1.nuorder.com for sandbox' = None,
@@ -147,30 +244,38 @@ def post(
     oauth_token_secret: 'OAuth token secret' = None,
     data: 'The data to send along with POST/PUT. If `-` then read from stdin.' = None,
     config_section: 'The name of the config section to get settings from.' = 'sandbox',
-    log_level: 'The log level to use.' = 'WARNING',
+    log_level: 'The log level to use.' = None,
     dry_run: "Don't actually run command, just show what would be run." = False,
     gzip_data: "Gzip data to NuOrder" = False
 ):
-    """Make a PUT request to NuOrder"""
+    """Make a POST request to NuOrder"""
     _set_log_level(log_level)
-    resp_json = nuorder.request(
-        method='POST',
-        endpoint=endpoint,
-        hostname=hostname,
-        consumer_key=consumer_key,
-        consumer_secret=consumer_secret,
-        oauth_token=oauth_token,
-        oauth_token_secret=oauth_token_secret,
-        data=data,
-        config_section=config_section,
+
+    c = functools.partial(ini_config.get, config_section)
+
+    if data == '-':
+        data = sys.stdin.buffer.read()
+
+    nu = nuorder.NuOrder(
+        hostname=hostname or c('hostname'),
+        consumer_key=consumer_key or c('consumer_key'),
+        consumer_secret=consumer_secret or c('consumer_secret'),
+        oauth_token=oauth_token or c('oauth_token'),
+        oauth_token_secret=oauth_token_secret or c('oauth_token_secret'),
+    )
+    resp_json = nu.post(
+        endpoint,
+        data,
         dry_run=dry_run,
         gzip_data=gzip_data,
     )
     return json.dumps(resp_json, indent=2)
 
 
-@argh.wrap_errors([nuorder.config.ConfigKeyMissing], processor=_failure)
+@argh.wrap_errors(DEFAULT_WRAPPED_ERRORS, processor=_failure)
 @argh.arg('-c', '--config-section')
+@argh.arg('-d', '--data')
+@argh.arg('--dry-run')
 def put(
     endpoint: 'The endpoint to interact with.',
     hostname: 'E.g. wholesale.sandbox1.nuorder.com for sandbox' = None,
@@ -180,31 +285,38 @@ def put(
     oauth_token_secret: 'OAuth token secret' = None,
     data: 'The data to send along with POST/PUT. If `-` then read from stdin.' = None,
     config_section: 'The name of the config section to get settings from.' = 'sandbox',
-    log_level: 'The log level to use.' = 'WARNING',
+    log_level: 'The log level to use.' = None,
     dry_run: "Don't actually run command, just show what would be run." = False,
     gzip_data: "Gzip data to NuOrder" = False
 ):
     """Make a PUT request to NuOrder"""
     _set_log_level(log_level)
-    resp_json = nuorder.request(
-        method='PUT',
-        endpoint=endpoint,
-        hostname=hostname,
-        consumer_key=consumer_key,
-        consumer_secret=consumer_secret,
-        oauth_token=oauth_token,
-        oauth_token_secret=oauth_token_secret,
-        data=data,
-        config_section=config_section,
+
+    c = functools.partial(ini_config.get, config_section)
+
+    if data == '-':
+        data = sys.stdin.buffer.read()
+
+    nu = nuorder.NuOrder(
+        hostname=hostname or c('hostname'),
+        consumer_key=consumer_key or c('consumer_key'),
+        consumer_secret=consumer_secret or c('consumer_secret'),
+        oauth_token=oauth_token or c('oauth_token'),
+        oauth_token_secret=oauth_token_secret or c('oauth_token_secret'),
+    )
+    resp_json = nu.put(
+        endpoint,
+        data,
         dry_run=dry_run,
         gzip_data=gzip_data,
     )
     return json.dumps(resp_json, indent=2)
 
 
-command_parser = argh.ArghParser(description=nuorder.__doc__)
+command_parser = argh.ArghParser(description=__doc__)
 command_parser.add_commands([
     initiate,
+    interact,
     get,
     delete,
     post,