+++ nuorder/_metadata.py
@@ -1,2 +1,4 @@
-__version_info__ = (1, 0, 1)
-__version__ = '.'.join(map(str, __version_info__))
+import pkg_resources
+
+__version__ = pkg_resources.require('nuorder')[0].version
+__version_info__ = tuple(int(p) for p in __version__.split('.'))